test2

+++ boot.sh
@@ -17,17 +17,26 @@ ansi_art='                 ▄▄▄
 clear
 echo -e "\n$ansi_art\n"
 
+# Install git early since it's needed when running boot.sh as a non-root user
 sudo pacman -Syu --noconfirm --needed git
 
-# Use custom repo if specified, otherwise default to basecamp/omarchy
-OMARCHY_REPO="${OMARCHY_REPO:-basecamp/omarchy}"
+OMARCHY_REPO="${OMARCHY_REPO:-basecamp/omarchy}" # custom repo with default fallback
+OMARCHY_REF="${OMARCHY_REF:-master}" # custom branch/ref with default fallback
+
+if [[ $EUID -eq 0 ]]; then
+  echo "Running as Root! Executing user-setup.sh to create non-root user and re-run boot.sh..."
+
+  curl -s "https://raw.githubusercontent.com/${OMARCHY_REPO}/${OMARCHY_REF}/install/bootstrap/user-setup.sh" | \
+    OMARCHY_REPO="${OMARCHY_REPO}" OMARCHY_REF="${OMARCHY_REF}" bash
+
+  # user-setup.sh will create user and re-run boot.sh as that user, then exit
+  exit 0 # exit to not run the rest of the script, and avoid cloning as root
+fi
 
 echo -e "\nCloning Omarchy from: https://github.com/${OMARCHY_REPO}.git"
 rm -rf ~/.local/share/omarchy/
 git clone "https://github.com/${OMARCHY_REPO}.git" ~/.local/share/omarchy >/dev/null
 
-# Use custom branch if instructed, otherwise default to master
-OMARCHY_REF="${OMARCHY_REF:-master}"
 if [[ $OMARCHY_REF != "master" ]]; then
   echo -e "\e[32mUsing branch: $OMARCHY_REF\e[0m"
   cd ~/.local/share/omarchy
